add log_record.hpp
