+++ sa/CMakeLists.txt
@@ -342,6 +342,7 @@ set(TRANSACTION_SOURCES
 set(TRANSACTION_HEADERS
   ${TRANSACTION_DIR}/log_common_impl.h
   ${TRANSACTION_DIR}/log_lsa.hpp
+  ${TRANSACTION_DIR}/log_record.hpp
   ${TRANSACTION_DIR}/log_system_tran.hpp
   )
